revert r42251. Cast the enum to uint32 before use in switchs to please GCC 4.5.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@42266 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/tracker/IconCache.cpp
@@ -647,7 +647,7 @@ IconCache::GetWellKnownIcon(AutoLock<SimpleIconCache> *,
 		// match up well known entries in the file system with specialized
 		// icons stored in Tracker's resources
 		int32 resid = -1;
-		switch (wellKnownEntry->which) {
+		switch ((uint32)wellKnownEntry->which) {
 			case B_BOOT_DISK:
 				resid = R_BootVolumeIcon;
 				break;
